+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Error when setting portal behavior on a new world
+
 ## [1.4.0] - 2026-01-04
 ### Added
 - World export command