+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.1] - 2026-01-31
+### Changed
+- Improved last location logic
+
 ### Fixed
 - Error when setting portal behavior on a new world